Make XMLRPC testing infrastructure reusable:
Move it from `.xmlrpc.tests` to `.xmlrpc.testing`.
It now requires the WSGI app to be provided.
Use the `testing` extra from `setup.py` to make use of this testing
infrastructure.
